+++ src/app.rs
@@ -242,13 +242,11 @@ impl App {
             debug!("Received key {:?}", key.code);
             match key.code {
                 KeyCode::Char('q') => app.mode = Mode::Quit,
-                KeyCode::Char('c') => app.content = Content::Countdown,
-                KeyCode::Char('t') => app.content = Content::Timer,
-                KeyCode::Char('p') => app.content = Content::Pomodoro,
-                // TODO(#102) Before we can use `e` here
-                // we do need to change keybindings for editing.
-                KeyCode::Char('z') => app.content = Content::Event,
-                KeyCode::Char('l') => app.content = Content::LocalTime,
+                KeyCode::Char('1') | KeyCode::Char('c') /* TODO: deprecated, remove it in next verson */ => app.content = Content::Countdown,
+                KeyCode::Char('2') | KeyCode::Char('t') /* TODO: deprecated, remove it in next verson */ => app.content = Content::Timer,
+                KeyCode::Char('3') | KeyCode::Char('p') /* TODO: deprecated, remove it in next verson */ => app.content = Content::Pomodoro,
+                KeyCode::Char('4') => app.content = Content::Event,
+                KeyCode::Char('0') | KeyCode::Char('l') /* TODO: deprecated, remove it in next verson */ => app.content = Content::LocalTime,
                 // toogle app time format
                 KeyCode::Char(':') => {
                     if app.content == Content::LocalTime {